Added new Type of System
Added a new Queue System that is not bound to entities.
Instead to only recieves entities in this queue this system can recieve
anything.
It breaks the paradigm that Entity System process entities.
You can have an entity system that deals with entities but does not
process one DIRECTELY

Added Tests to all platforms
Added changes to all platfotms