Improved MultiThreading
No more ExecutionType parameter on Update and Draw.
ExecutionType is now defined by system (we can have in the same world
systems that are regular and others that are multithreaded)
Artemis goes for each layer and process the regular ones in a sync way,
then it create some tasks to process the async ones of the same layer.
Then it procedes to the next layer ....